test(downloader): Split `forUrl()` to `funTest`
These actually depend on the `git` executable and thus are not pure
(unit) tests.
