+++ downloader/src/test/kotlin/VersionControlSystemTest.kt
@@ -20,13 +20,8 @@
 package org.ossreviewtoolkit.downloader
 
 import io.kotest.core.spec.style.WordSpec
-import io.kotest.matchers.nulls.beNull
-import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.result.shouldBeSuccess
-import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
-import io.kotest.matchers.shouldNot
-import io.kotest.matchers.shouldNotBe
 
 import io.mockk.every
 import io.mockk.mockk
@@ -121,45 +116,6 @@ class VersionControlSystemTest : WordSpec({
                 "master",
                 "main"
             )
-        }
-    }
-
-    "forUrl()" should {
-        "return null for an unsupported repository URL" {
-            val repositoryUrl = "https://example.com"
-
-            val vcs = VersionControlSystem.forUrl(repositoryUrl)
-
-            vcs should beNull()
-        }
-
-        "return a VCS instance that can handle a Git repository URL" {
-            val repositoryUrl = "https://github.com/oss-review-toolkit/ort.git"
-
-            val vcs = VersionControlSystem.forUrl(repositoryUrl)
-
-            vcs shouldNotBeNull {
-                type shouldBe VcsType.GIT
-            }
-        }
-
-        "return the VCS instance with the correct configuration from cache" {
-            val repositoryUrl = "https://github.com/oss-review-toolkit/ort.git"
-            val configs = mapOf(
-                VcsType.GIT.toString() to PluginConfig(
-                    options = mapOf("updateNestedSubmodules" to false.toString())
-                )
-            )
-
-            val vcsWithDefaultConfiguration = VersionControlSystem.forUrl(repositoryUrl)
-            val vcsWithConfigs = VersionControlSystem.forUrl(repositoryUrl, configs)
-            val vcsWithConfigsFromCache = VersionControlSystem.forUrl(repositoryUrl, configs)
-
-            vcsWithDefaultConfiguration shouldNot beNull()
-            vcsWithConfigs shouldNot beNull()
-
-            vcsWithDefaultConfiguration shouldNotBe vcsWithConfigs
-            vcsWithConfigsFromCache shouldBe vcsWithConfigs
         }
     }
 })
